Update django.po
When testing the MRI version of the rhythm experiment, I found that some of the Dutch translations of the consent form were missing. One of the translations ("Anybody aged 16 or older with no hearing problems is (....) use either internal or external loudspeakers.") was already in the .po file but somehow does not appear in the consent form. Could you maybe check why that is?