story: Add MatchingPairs story with default props and two columns support
This commit adds the MatchingPairs component with its default props and support for displaying two columns of sections. The component automatically adjusts the number of columns based on the number of sections provided. If there are six or less sections, two columns are displayed. If there are more than six sections, four columns are displayed. The component is also displayed in fullscreen mode.