feat: Update e2e run-tests script to support interactive mode
The e2e run-tests script has been updated to support an interactive mode. Now, when the script is run with the '-i' parameter, it starts the container with an interactive bash shell. If no '-i' parameter is supplied, the script runs the python test script. This change improves the flexibility and usability of the script for running Selenium tests in a Docker container.