refactor: Update e2e run-tests script to pass BASE_URL as environment variable
The e2e run-tests script has been refactored to pass the BASE_URL as an environment variable to the Docker container. This change improves the flexibility and maintainability of the script by decoupling the base URL configuration from the script itself. Now, the script can easily adapt to different environments without modifying the code.