Fix (CI): Fix frontend deployment (#1011)
* Fix (CI): Fix frontend build in deployment to tst & acc (#1007)

* config: Set SENTRY_ENVIRONMENT to "test" and "acceptance" in relevant files so Sentry knows which environment it is running on (#972)

* ci: Update podman.yml workflow conditions

The `if` conditions in the `podman.yml` workflow file have been updated to include additional checks for the `workflow_dispatch` event. This ensures that the workflow is triggered correctly for the workflow dispatch button in combination with either the `develop` or the `main` branch.

* chore: Update podman.yml workflow variables to avoid the vite build failing due to missing environment variables

This commit updates the `podman.yml` workflow file to include additional variables related to the frontend HTML. These variables are used for the favicon, logo URL, Open Graph (OG) description, OG image, OG title, OG URL, and body class. The variables are set to empty strings if not provided. This change ensures that the workflow has the necessary variables for the frontend HTML.

* ci: Temporarily turn on deploy for acceptance on this branch

* chore: Try if manually setting the favicon fixes things

* ci: Update podman.yml workflow variables to include frontend HTML environment variables

This commit updates the `podman.yml` workflow file to include additional variables related to the frontend HTML. These variables are used for the favicon, logo URL, Open Graph (OG) description, OG image, OG title, OG URL, and body class. The variables are set to empty strings if not provided. This change ensures that the workflow has the necessary variables for the frontend HTML.

* ci: Test deploy to test environment

* chore: Update podman.yml workflow conditions for develop branch

(cherry picked from commit 257b3a7514b51170f2235f0eb6035264e0a780c9)

* chore: Update podman.yml workflow conditions for acceptance deployment

* fix: Preload audio without MEDIA_ROOT

* revert: Do not automatically deploy this PR's commits to acceptance anymore as I (Drikus) have confirmed that Acceptance works again

* revert: Revert fix as I will address it in a separate PR