Chore(deps): Remove IPToCC and numpy from base requirements (#1161)
* chore(deps): Remove IPToCC from base requirements

The commit removes the `IPToCC` package from the base requirements file. This package was used to get the country code from an IP address, but it is no longer needed and caused 15m Docker image build times

* chore(deps): (Re-)install numpy as we are using it.

This was an indirect dependency of iptocc and got removed when iptocc was removed from the pythonic requirements.

* chore: Replace numpy import to use random module

The code changes replace the `numpy` import with the `random` module in the `speech2song.py` and `practice.py` files. This change improves performance and removes the need for the `numpy` dependency.