Added: Add SponsorImage model and update FooterConfig model (#1188)
* refactor: Add SponsorImage model and update FooterConfig model

This commit adds a new model called SponsorImage and updates the FooterConfig model in the backend/theme app. The SponsorImage model represents an image associated with a footer configuration, and it includes fields for the index, footer configuration, and image. The FooterConfig model now uses the SponsorImage model through a many-to-many relationship for the logos field. This change allows multiple images to be associated with a footer configuration.

* feat: Add nested inline form for sponsor image on footer config on theme config

* fix: Fix reverse migration

* fix(tests): Fix tests related to footer logos, which now have a SponsorImage through model and should also be tested for correct sorting using the index field

* fix(lint): Fix linting errors in migration file

* feat: Add drag & drop sorting!